+++ src/shared/layout/HeaderLessLayout.tsx
@@ -1,0 +1,4 @@
+function HeaderLessLayout() {
+  return <div>HeaderLessLayout</div>;
+}
+export default HeaderLessLayout;
