+++ src/shared/layout/PageLayout.tsx
@@ -1,0 +1,4 @@
+function PageLayout() {
+  return <div>PageLayout</div>;
+}
+export default PageLayout;
